fix: Improve PBS error reporting [FOUNDENG-248] (#453)
Augment the error logs with the HTTP response value on failure.  The returned
error does not always have the underlying info (e.g. 404 Not Found).

Fix the patterns use for matching messages with PBS.   Add an entry for
Slurm (which was showing up previously because no messages matched).

Add filtering based up on the reporter to eliminate some noise that we never
want to see from the Dispatcher infrastructure.