fix: Restore det shell on podman [FOUNDENG-280] (#493)
When running rootless podman, inside the container we are
root/uid=0 and that maps to the user account outside the
container.   All is fine until we attempt to ssh into the
container which actually then uses the launching username/uid.
Under normal circumstance /run/determined/ssh has only 0600
permissions for only the owning user to read, but with podman
root maps to the user/uid, so the launching user is not seen
as having access to the files.

Until we find a better solution, dynamically relax the permissions
to be a+x on the /run/determined/ssh directory path such that
the user can read /run/determined/ssh/authorized_keys and enable
ssh into the container to work proplerly.

Additionally, drop use of the podman --hostuser arg, as it doens't help
the situation and we already provide the launching user in a custom
passwd entry.