feat: Fully support apptainer fork of singularity [FOUNDENG-292] (#507)
Apptainer 1.0 is a fork of Singularity 3.8.  Reduce use of SINGULARITY_* variables.
hpe-hpc-launcher 3.1.4 supports capabilities and cached bypass.   --no-mount=tmp
has been the default for a bit, so not explicitly needed.

We retain the use of the SINGULARITY_DOCKER* and add APPTAINER_DOCKER*
for creds as there is no CLI option alternative.   Adding the APPTAINER_* version
eliminates warnings.