chore: Add ENROOT_RUNTIME_PATH to environment for enroot (#578)
This change adds ENROOT_RUNTIME_PATH=/tmp/$(whoami) to
the job batch script to enable enroot testing without additional
configuration.  The whoami is deferred until batch job execution
to get the value from the running HPC cluster user.   It is set
for all casablanca nodes regardless of container type, but is
ignored in the other cases.

Defining ENROOT_RUNTIME_PATH in the batch job does not currently
work for PBS (casablanca-login2) since the variables are not inherited
to the compute nodes (and instead is passed directly into the container).