fix: Enable IB transport and MPI images to work by default [FOUNDENG-412] (#628)
The default "max locked memory" in our containers is 64, which is too low for IB
If we are getting the default value in the container, then raise it to
unlimited.   If it is set to some other value, leave it as-is to
allow a customer override.

If no user SLURM_MPI_TYPE is defined, set it to pmi2 as needed by the
Determined MPI environment image.