fix(dispatcherrm): ResolveResourcePool impl is semantically wrong (#799) [DET-9286]
In OSS Determined, a fix was landed so that we resolve task container defaults with the
correct resource pool but, because DispatcherRM's ResolveResourcePool impl was slightly
wrong, we were accidentally relying on the incorrect behavior to end up with the right
defaults. The callers of ResolveResourcePool do no expect to receive 'providing partition'
but just to have defaults applied, because of this we got the wrong task container defaults
and it caused the runtime expconf contain the wrong pool.